requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.2.0
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
